==> Notebooks/01_scripts_R/04_decission_tree.Rmd <==
---
title: "Decission Tree"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

```{r cars}
library(rpart)
library(rpart.plot)
library(readr)
library(dplyr)

  # Cargamos los datos

data <- read_csv("../00_data/3_clusters/data_clustering.csv")

data_filter <- data %>%
  select(-X1, -X1_1, -cod_cpi, -crash_categorica, -n_crashes, -dias_desde_ultimo_arranque) %>%
  mutate(
         sexo = as.factor(sexo),
         gpos_aut = as.factor(gpos_aut),
         Grupos_variables = as.factor(Grupos_variables),
         tipo_de_marca = as.factor(tipo_de_marca)) 

  # Crecimiento de arbol
modelo <- rpart(crash_flag ~ ., data = data_filter)
modelo

  # Plot
rpart.plot(x = modelo)

```

## Including Plots

You can also embed plots, for example:

```{r pressure, echo=FALSE}
plot(pressure)
```

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
